carlib: regenerate IEventService.aidl from decompiled interface (144 methods, exact transaction ordinals) + AIDL_ORDINALS.md
Unblocks radio/climate/volume/EQ widgets: transact() codes now match the device.
ICameraService stub + Surface import added; CarService.getMainVolume corrected to byte->Int.

+++ launcher/carlib/src/main/java/com/reveng/carlauncher/carlib/CarService.kt
@@ -100,7 +100,7 @@ class CarService(private val appContext: Context) {
     fun getValidMode(): Int? = call { getValidMode() }
     fun isBackCarConnected(): Boolean = call { IsBackCarConneted() } ?: false
     fun getRadioFreq(): Int? = call { getRadioFreq() }
-    fun getMainVolume(): Int? = call { getMainVolval() }
+    fun getMainVolume(): Int? = call { getMainVolval().toInt() }
     fun getMcuVer(): String? = call { getCanVer() }
 
     fun sendMode(mode: Int, flag: Boolean) { call { sendMode(mode, flag) } }
